requirements.txt
discord.py==1.7.3
aiohttp
requests
colorama
Flask
pystyle
bs4
pyfiglet==0.8.post1
openai==0.28.0
python-dateutil==2.8.2
googletrans==4.0.0-rc1
qrcode==7.3
image
discord_interactions
spotipy
ffmpeg-python
pystyle
pytz
python-chess
speedtest-cli
requests
pyttsx3
wikipedia
beautifulsoup4
Pillow
python-dotenv
aiohttp
pycryptodome
base58
mnemonic
ecdsa
setuptools